wip: refactor/app.js cleanup
