+++ elixir/web-ng/test_websock.exs
@@ -3,4 +3,5 @@ defmodule TestWebSock do
     IO.puts("WebSockAdapter exists? #{Code.ensure_loaded?(WebSockAdapter)}")
   end
 end
+
 TestWebSock.run()
